# api-gateway/docs/swagger.yaml
definitions:
  domain.PaginatedResponse:
    properties:
      data: {}
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  domain.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - full_name
    - password
    type: object
  handlers.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
  description: Hệ thống xử lý tài liệu thông minh (Intelligent Document Processing).
  title: IDP API Gateway
  version: "1.0"
paths:
  /api/v1/admin/jobs:
    get:
      description: Returns all jobs with pagination, optional status/file_code filters,
        and associated user/document info. ADMIN only.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by file code
        in: query
        name: file_code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.PaginatedResponse'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get All Jobs (Paginated)
      tags:
      - admin
  /api/v1/admin/stats:
    get:
      description: Returns total users, total jobs, and job counts grouped by state.
        ADMIN only.
      produces:
      - application/json
      responses:
        "200":
          description: System statistics
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get System Statistics
      tags:
      - admin
  /api/v1/admin/users:
    get:
      description: Returns all registered users (passwords excluded). ADMIN only.
      produces:
      - application/json
      responses:
        "200":
          description: List of all users
          schema:
            items:
              $ref: '#/definitions/domain.User'
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get All Users
      tags:
      - admin
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and set HttpOnly cookies (access_token 15m +
        refresh_token 7d).
      parameters:
      - description: Login Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      description: Logout user, invalidate refresh token in Redis, and clear all auth
        cookies.
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Logout
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      description: Use refresh_token cookie to obtain a new short-lived access_token.
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid or expired refresh token
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh Access Token
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email, password, and full name.
      parameters:
      - description: Register Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register new user
      tags:
      - auth
  /api/v1/jobs:
    get:
      description: Returns the authenticated user's jobs with pagination and optional
        filters.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: Filter by status (PENDING, EXTRACTING, COMPLETED, FAILED)
        in: query
        name: status
        type: string
      - description: Filter by file code (substring match)
        in: query
        name: file_code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get User's Jobs (Paginated)
      tags:
      - jobs
  /api/v1/jobs/{id}:
    get:
      parameters:
      - description: Job ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Job not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Job Status
      tags:
      - jobs
  /api/v1/jobs/{id}/stream:
    get:
      description: Subscribe to Server-Sent Events for real-time job updates.
      parameters:
      - description: Job ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: Event stream
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Job not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Stream Job Status (SSE)
      tags:
      - jobs
  /api/v1/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a PDF or Image file with metadata. Requires Bearer Token.
        Validates file content via Magic Bytes.
      parameters:
      - description: Document file (PDF, PNG, JPG)
        in: formData
        name: file
        required: true
        type: file
      - description: User-defined document code (alphanumeric, dashes, max 50 chars)
        in: formData
        name: file_code
        required: true
        type: string
      - description: Optional notes about the document
        in: formData
        name: notes
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns job_id and status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error or unsupported file type
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload Document
      tags:
      - jobs
  /api/v1/users/me:
    get:
      consumes:
      - application/json
      description: Get the profile of the currently authenticated user using JWT cookie
        or Bearer token fallback.
      produces:
      - application/json
      responses:
        "200":
          description: Returns user details without password
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
  /internal/webhook/job-completed:
    post:
      consumes:
      - application/json
      description: Called by the Python worker after updating a job to COMPLETED/FAILED.
        Triggers cache invalidation. Requires X-Webhook-Secret header.
      parameters:
      - description: Shared webhook secret
        in: header
        name: X-Webhook-Secret
        required: true
        type: string
      - description: 'Payload: {\'
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Cache invalidated
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Missing job_id
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Invalidation failed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 'Internal Webhook: Notify Job Completion'
      tags:
      - internal
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"